AI settleShip WIP
-Add a cursor attribute , and its getter/setter to the Player. This 
attribute will be the first coordinate of the settled ship
-Add a symbol attribute to the ship, which is the symbol what will be 
drawn on the Board if the settle is succesful.
-Add an emptyCoordinates attribute to the Board which containes the 
empty coordinates. So if we settle a ship, we add an  aura to it, the 
both of their coordinates will be removed form the empty coordinates 
list. If the ship coordinate is not in the empty coordinate list, that
means that one of the coordinates are occupied by another ship, or the
Board dimension is too small, so it not contains it. Whit this mechanic
we doesn't need individual methods to check that if the ship is out of
Board, or the ship is collosing with another ship.
-In the Board constructor we initialise the emptyCoordinates, so if we
get a premade Board,we create a List with all of the empty coordinates,
and we search for the occupied coordinates of it and adding an aura to 
it, and remove these coordinates too from the earlier defined emptyCoordinates list. If we got only the dimenions, we initilaise the
battlefield with empty coordinates, and we dont have an occupied
coordinate, so the aura dosen't matter.
-Write a test to check that the shipSymbol attribute in the Ship constructor not equal to EMPTY_COORDINATE constant or not NULL.
-Rewrite the settle test, with the new settle method which use the empty
coordinate system.